added support for additonal as well as multiple simultaneous -f arguments

+++ vidos_av1/vobu.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 SUPPORTED_VID_CODECS=( "av1" "vp8" "vp9" )
 STYLE_ARRAY=( "disk" "ram" "hybrid" )
-FIRMWARE_ARRAY=( "amdgpu" "radeon" "i915" )
+FIRMWARE_ARRAY=( "amdgpu" "radeon" "i915" "none" "all")
 VID_SUPPORTED=1
 ARG_VALID=0
 STYLE_VALID=1
-FIRMWARE_VALID=1
+FIRMWARE_SELECTION=()
 
 print_help() {
 echo "VidOS build ultilty
@@ -14,13 +14,12 @@ options:
 -h help -- print this help text
 -d directory -- path to iso filesystem root directory
 -v filename -- path to video file, supported video codecs: "${SUPPORTED_VID_CODECS[@]}"
--s build style -- style of output build: "${STYLE_ARRAY[@]}"
--f firmware -- binary graphics drivers: "${FIRMWARE_ARRAY[@]}""
+-s build style -- style of output build, one of: "${STYLE_ARRAY[@]}"
+-f firmware -- binary graphics drivers, one or multiple of: "${FIRMWARE_ARRAY[@]}""
 }
 
 checkArg() {
-	if [ $3 ]; then
-		echo $2 "is" $3
+	if [ $3 ]; then echo $2 "is" $3
 	fi
 	if [ $1 -ne 0 ]; then
 		echo $3 "is not supported as an option for "$2"!"
@@ -39,8 +38,7 @@ checkPaths() {
 }
 
 checkOpts() {
-	if [ -z $1 ] || [ "-" = $1 ] ; then
-	        echo "invalid option"
+	if [ -z $1 ] || [ "-" = $1 ] ; then echo "invalid option"
 	        ARG_VALID=1
 	fi
 }
@@ -66,6 +64,7 @@ while getopts ":h:d:v:s:f:" opt; do
 		f)
 			FIRMWARE="$OPTARG"
 			checkOpts $FIRMWARE "firmware"
+			FIRMWARE_SELECTION+=($FIRMWARE)
 		;;
 		*)
 			echo "Either empty or invalid option: $OPTARG"
@@ -79,6 +78,8 @@ checkArg $ARG_VALID "argument"
 
 VIDOS_ISO9660=$(ls $DIR | grep -m 1 "vidos_iso9660_*")
 
+mkdir -p $DIR/initramfs_overlay/firmware
+
 disk_VID_PATH=$DIR/$VIDOS_ISO9660/video/
 ram_VID_PATH=$DIR/initramfs_overlay/opt/
 disk_SED_ARG='s/^/\/media\/video\//'
@@ -90,23 +91,17 @@ AUD_CODEC=$(ffprobe -v quiet -show_streams $VIDEO | grep -w codec_name | sed -n 
 
 for CODEC in "${SUPPORTED_VID_CODECS[@]}"
 do
-	if [ $VID_CODEC = $CODEC ]; then
-		VID_SUPPORTED=0
-		VID_FORMAT=$VID_CODEC
-		if [[ $VID_CODEC = vp[8-9] ]]; then
-			VID_FORMAT="webm"
+	if [ $VID_CODEC = $CODEC ]; then VID_SUPPORTED=0 VID_FORMAT=$VID_CODEC
+		if [[ $VID_CODEC = vp[8-9] ]]; then VID_FORMAT="webm"
 		fi
 	fi
 done
 
-if [ $AUD_CODEC = "opus" ]; then
-	echo "audio passed"
-else
-	echo "Warning: audio is" $AUD_CODEC", not opus, so it will not play back!"
+if [ $AUD_CODEC = "opus" ]; then echo "audio passed"; else
+echo "Warning: audio is" $AUD_CODEC", not opus, so it will not play back!"
 fi
 
-if [ -z $VIDOS_ISO9660 ]; then
-	echo "vidos iso9660 root filesystem not found."
+if [ -z $VIDOS_ISO9660 ]; then echo "vidos iso9660 root filesystem not found."
 	exit 1
 elif [ $VIDOS_ISO9660 != "vidos_iso9660_"$VID_FORMAT ] && [ $VID_SUPPORTED = 0 ]; then
 	echo "video is not of correct type for currently built VidOS"
@@ -115,7 +110,6 @@ elif [ $VIDOS_ISO9660 != "vidos_iso9660_"$VID_FORMAT ] && [ $VID_SUPPORTED = 0 ]
 fi
 
 checkArg $VID_SUPPORTED "video" $VID_CODEC "${SUPPORTED_VID_CODECS[*]}"
-
 
 for STYLE_OPTION in "${STYLE_ARRAY[@]}"
 do
@@ -144,33 +138,33 @@ done
 
 checkArg $STYLE_VALID "style" $STYLE
 
-for FIRMWARE_OPTION in "${FIRMWARE_ARRAY[@]}"
+for i in "${FIRMWARE_SELECTION[@]}"
 do
-        if [ $FIRMWARE = $FIRMWARE_OPTION ]; then
-		FIRMWARE_VALID=0
-		cp -r $DIR/firmware/$FIRMWARE $DIR/initramfs_overlay/firmware
-        fi
+	FIRMWARE_VALID=1
+	for FIRMWARE_OPTION in "${FIRMWARE_ARRAY[@]}"
+	do
+	        if [ $i = $FIRMWARE_OPTION ]; then
+			FIRMWARE_VALID=0
+			FIRMWARES=$i
+			if [ -h $DIR/firmware/$i ]; then FIRMWARES=$i/*; fi
+			cp -r $DIR/firmware/$FIRMWARES -t $DIR/initramfs_overlay/firmware/
+	        fi
+	done
+			checkArg $FIRMWARE_VALID "firmware" $i "${FIRMWARE_ARRAY[*]}"
 done
 
-checkArg $FIRMWARE_VALID "firmware" $FIRMWARE "${FIRMWARE_ARRAY[*]}"
+cd $DIR/initramfs_overlay
+find . | LC_ALL=C sort | cpio --quiet -o -H  newc > ../$VIDOS_ISO9660/kernel/rootfs.cpio
+cd ../
+lz4 -l -9 -c $VIDOS_ISO9660/kernel/rootfs.cpio > $VIDOS_ISO9660/kernel/rootfs.cpio.lz4
+rm $VIDOS_ISO9660/kernel/rootfs.cpio
+echo "installed "$VIDEO" as a new video and rebuilt playlist"
+echo "rebuilding iso"
+xorriso -as mkisofs -quiet -o ../vidos_$VID_CODEC"_""$(IFS=_ ; echo "${FIRMWARE_SELECTION[*]}")""_"$VIDEO_NAME.iso -isohybrid-mbr isohdpfx.bin \
+-c isolinux/boot.cat -b isolinux/isolinux.bin -no-emul-boot -boot-load-size 4 -boot-info-table \
+$VIDOS_ISO9660
 
-	cd $DIR/initramfs_overlay
-	find . | LC_ALL=C sort | cpio --quiet -o -H  newc > ../$VIDOS_ISO9660/kernel/rootfs.cpio
-	cd ../
-	lz4 -l -9 -c $VIDOS_ISO9660/kernel/rootfs.cpio > $VIDOS_ISO9660/kernel/rootfs.cpio.lz4
-	rm $VIDOS_ISO9660/kernel/rootfs.cpio
-	echo "installed "$VIDEO" as a new video and rebuilt playlist"
-	echo "rebuilding iso"
-	xorriso -as mkisofs -quiet -o ../vidos_$VID_CODEC"_"$FIRMWARE"_"$VIDEO_NAME.iso -isohybrid-mbr isohdpfx.bin \
-	-c isolinux/boot.cat -b isolinux/isolinux.bin -no-emul-boot -boot-load-size 4 -boot-info-table \
-	$VIDOS_ISO9660
-
-	if [ -n "$(ls $VIDOS_ISO9660/video/)" ]; then
-		rm $VIDOS_ISO9660/video/*
-	elif [ -n "$(ls initramfs_overlay/opt/)" ]; then
-		rm initramfs_overlay/opt/*
-	elif [ -n "$(ls initramfs_overlay/firmware/)" ]; then
-		rm initramfs_overlay/firmware/*
-	fi
-
-
+rm -r initramfs_overlay/firmware/
+if [ -n "$(ls $VIDOS_ISO9660/video/)" ]; then rm $VIDOS_ISO9660/video/*
+elif [ -n "$(ls initramfs_overlay/opt/)" ]; then rm initramfs_overlay/opt/*
+fi
